fix: 修复 M02/M03 模块 EF Core 模型与数据库列名映射不一致问题
- M02 基础数据: 10 个 Controller 的 int id 参数类型统一改为 long id
- M02 模型: Equipment 表名修正(equipments→equipment), Product 添加 OrgId,
  DynamicParam/ParamRealtimeValue/InspectionPlan/InspectionPlanItem
  补全 [Column] 属性映射
- M03 IQC: IqcReceipt/IqcInspection/IqcInspectionItem/IqcAnomaly
  补全 [Column] 属性映射, SupplierScore 修正字段映射, 移除导航属性错误 [Column]
- Program.cs: 更新所有 CREATE TABLE SQL 语句, 补全 is_active/created_at/updated_at 列
- 前端: 修复 audit/equipmentLink 类型定义, SPC 参数映射,
  46 个功能页 UI 细节优化, IQC/IPQC 前端页面适配新 API 结构

+++ backend/Controllers/M02/BomsController.cs
@@ -38,7 +38,7 @@ public class BomsController : ControllerBase
     }
 
     [HttpGet("{id}")]
-    public async Task<ActionResult<Bom>> Get(int id)
+    public async Task<ActionResult<Bom>> Get(long id)
     {
         var entity = await _db.Boms.Include(b => b.Product).FirstOrDefaultAsync(b => b.Id == id);
         if (entity == null) return NotFound();
@@ -63,7 +63,7 @@ public class BomsController : ControllerBase
     }
 
     [HttpPut("{id}")]
-    public async Task<ActionResult<Bom>> Update(int id, [FromBody] UpdateBomDto dto)
+    public async Task<ActionResult<Bom>> Update(long id, [FromBody] UpdateBomDto dto)
     {
         var entity = await _db.Boms.FindAsync(id);
         if (entity == null) return NotFound();
@@ -79,7 +79,7 @@ public class BomsController : ControllerBase
     }
 
     [HttpDelete("{id}")]
-    public async Task<IActionResult> Delete(int id)
+    public async Task<IActionResult> Delete(long id)
     {
         var entity = await _db.Boms.FindAsync(id);
         if (entity == null) return NotFound();
